Fix: 회원가입 승인 모달 제거#132
